move fluorescence plate block to softmax pro structure
